layouts

+++ src/app/service/token-check.service.ts
@@ -1,5 +1,6 @@
+import { DOCUMENT } from '@angular/common';
 import { HttpErrorResponse, HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
-import { Injectable } from '@angular/core';
+import { Inject, Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 import { catchError, Observable, throwError } from 'rxjs';
 
@@ -10,19 +11,24 @@ export class TokenCheckService implements HttpInterceptor {
 
   constructor(
     private router: Router,
-
+    @Inject(DOCUMENT) private document: Document
   ) { }
 
   intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-    let token = null;
-    if (typeof window !== 'undefined' && localStorage) { // 確保只有在瀏覽器中執行
-      token = localStorage.getItem("token");
+    // let token = null;
+    // if (typeof window !== 'undefined' && localStorage) { // 確保只有在瀏覽器中執行
+    //   token = localStorage.getItem("token");
+    // }
+    const localStorage = document.defaultView?.localStorage;
+    if (localStorage) {
+      const token = localStorage.getItem('token');
+      if (token) {
+        req = req.clone({
+          setHeaders: { Authorization: `Bearer ${token}` }
+        });
+      }
     }
-    if (token) {
-      req = req.clone({
-        setHeaders: { Authorization: `Bearer ${token}` }
-      });
-    }
+
 
     return next.handle(req).pipe(
       catchError((err) => {
@@ -32,7 +38,9 @@ export class TokenCheckService implements HttpInterceptor {
             if (this.router.url === '/') {
 
             } else {
-              localStorage.clear();
+              if (localStorage) {
+                localStorage.clear();
+              }
               this.router.navigateByUrl('/');
             }
           }
